+++ app/entity/response/SearchBannerResponse.go
@@ -1,7 +1,7 @@
 package response
 
 type SearchBannerResponse struct {
-	BannerId uint                `json:"banner_id"`
+	BannerId int64               `json:"banner_id"`
 	Image    string              `json:"image"`
 	Status   int                 `json:"status"`
 	Movie    SearchMovieResponse `json:"movie"`
